change files layout add suffix and prefix for Thymeleaf replace @Autowired with @Overwrite on configureGlobal method in SecurityConfig and change it's name to configure
