--- src/main/java/ua/goit/group2notes/ServletInitializer.java
@@ -1,13 +1,0 @@
-package ua.goit.group2notes;
-
-import org.springframework.boot.builder.SpringApplicationBuilder;
-import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
-
-public class ServletInitializer extends SpringBootServletInitializer {
-
-	@Override
-	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
-		return application.sources(Group2NotesApplication.class);
-	}
-
-}
