+++ src/main/java/ua/goit/group2notes/errorHandling/UserAlreadyExistsException.java
@@ -1,4 +1,4 @@
-package ua.goit.errorHandling;
+package ua.goit.group2notes.errorHandling;
 
 public class UserAlreadyExistsException extends RuntimeException {
     public UserAlreadyExistsException(String message) {
