+++ src/main/java/ua/goit/group2notes/user/UserDto.java
@@ -1,25 +1,22 @@
-package ua.goit.user;
+package ua.goit.group2notes.user;
 
-import ua.goit.note.NoteDao;
+import ua.goit.group2notes.note.NoteDto;
 
-import javax.persistence.*;
+import javax.validation.constraints.Pattern;
+import javax.validation.constraints.Size;
 import java.util.Set;
 import java.util.UUID;
 
-@Entity
-@Table(name = "users")
-public class UserDao {
+public class UserDto {
     private UUID id;
     private String username;
     private String password;
     private UserRole userRole;
-    private Set<NoteDao> notes;
+    private Set<NoteDto> notes;
 
-    public UserDao() {
+    public UserDto() {
     }
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
     public UUID getId() {
         return id;
     }
@@ -28,7 +25,9 @@ public class UserDao {
         this.id = id;
     }
 
-    @Column(name = "username", nullable = false, unique = true)
+//    @NotEmpty(message = "Please, enter Username")
+    @Size(min = 5, max = 50, message = "Username must be minimum 5 characters long and no longer then 50")
+    @Pattern(regexp = "[a-zA-Z0-9]*", message = "Username must contains only letters or/and digits")
     public String getUsername() {
         return username;
     }
@@ -37,7 +36,8 @@ public class UserDao {
         this.username = username;
     }
 
-    @Column(name = "password", nullable = false)
+//    @NotEmpty(message = "Please, enter Password")
+    @Size(min = 8, max = 100, message = "Password must be minimum 8 characters long and no longer then 100")
     public String getPassword() {
         return password;
     }
@@ -46,8 +46,6 @@ public class UserDao {
         this.password = password;
     }
 
-    @Column(name = "user_role", nullable = false)
-    @Enumerated(EnumType.STRING)
     public UserRole getUserRole() {
         return userRole;
     }
@@ -56,12 +54,11 @@ public class UserDao {
         this.userRole = userRole;
     }
 
-    @OneToMany(mappedBy = "user", fetch = FetchType.EAGER)
-    public Set<NoteDao> getNotes() {
+    public Set<NoteDto> getNotes() {
         return notes;
     }
 
-    public void setNotes(Set<NoteDao> notes) {
+    public void setNotes(Set<NoteDto> notes) {
         this.notes = notes;
     }
 }
