+++ src/main/resources/application.properties
@@ -7,4 +7,7 @@ server.port=9999
 
 spring.flyway.locations=classpath:db/migration/h2
 
+spring.thymeleaf.prefix=classpath:/templates/
+spring.thymeleaf.suffix=.html
+
